Updated build.gradle
Commented out googleJavaFormat() in spotless section.

I was seeing crashes of the spotless tool during builds from this module which was preventing a successful build.  I wasn't completely crazy about how it reformatted code but the crashing was the real issue - any optional tool that causes a build to fail because the *TOOL* itself fails should be tossed out because it's broken.